fix(cpu/matmul): widen the sibling 4-row GEMV offsets too (#429)
Completes the previous port. The first pass widened only
`weightsQ8 + row * rowBytes` and left the three sibling pointers of the
4-row unrolled VNNI/AVX-512 calls (`weightsQ8 + (row + 1|2|3) * rowBytes`)
in 32-bit arithmetic, in both MatMul.cs and the fork-only MatMulVnni.cs.
Those overflow on the same tensors and slightly sooner, since they address
further into the matrix, so the hot path the fix was meant to protect was
still exposed.

Also widens GemvF32's weight row offsets (`a + row * k`); `a` is documented
as "weight matrix A [M×K]", so it is the same class as the GemvF16 sites.
Needs >2^31 float elements (~8.6 GB) to wrap, so consistency rather than a
reachable defect.

MatMulMxfp4.cs needs no change: its GEMV already declares
`long rowBytes = (long)blockCount * Mxfp4BlockBytes`, so the multiply is
already 64-bit.

Caught by Copilot review on kkokosa/dotLLM#431.

+++ src/DotLLM.Cpu/Kernels/MatMul.cs
@@ -58,7 +58,7 @@ public static unsafe partial class MatMul
 
         for (int row = 0; row < m; row++)
         {
-            var rowSpan = new ReadOnlySpan<float>(a + row * k, k);
+            var rowSpan = new ReadOnlySpan<float>(a + (long)row * k, k);
             result[row] = TensorPrimitives.Dot(rowSpan, xSpan);
         }
     }
@@ -72,7 +72,7 @@ public static unsafe partial class MatMul
         for (int row = 0; row < m; row++)
         {
             float sum = 0;
-            float* rowPtr = a + row * k;
+            float* rowPtr = a + (long)row * k;
             for (int j = 0; j < k; j++)
                 sum += rowPtr[j] * x[j];
             result[row] = sum;
@@ -145,9 +145,9 @@ public static unsafe partial class MatMul
             {
                 VecDotQ8_0Avx512_4Rows(
                     weightsQ8 + (long)row * rowBytes,
-                    weightsQ8 + (row + 1) * rowBytes,
-                    weightsQ8 + (row + 2) * rowBytes,
-                    weightsQ8 + (row + 3) * rowBytes,
+                    weightsQ8 + (long)(row + 1) * rowBytes,
+                    weightsQ8 + (long)(row + 2) * rowBytes,
+                    weightsQ8 + (long)(row + 3) * rowBytes,
                     xQ8, blockCount, result + row);
             }
             for (; row < m; row++)
@@ -163,9 +163,9 @@ public static unsafe partial class MatMul
             {
                 VecDotQ8_0Avx2_4Rows(
                     weightsQ8 + (long)row * rowBytes,
-                    weightsQ8 + (row + 1) * rowBytes,
-                    weightsQ8 + (row + 2) * rowBytes,
-                    weightsQ8 + (row + 3) * rowBytes,
+                    weightsQ8 + (long)(row + 1) * rowBytes,
+                    weightsQ8 + (long)(row + 2) * rowBytes,
+                    weightsQ8 + (long)(row + 3) * rowBytes,
                     xQ8, blockCount, result + row);
             }
             for (; row < m; row++)
